fixed i1 and i2 data type (made uint16_t)

+++ v1.1/sketch_esp_WiFi_UART_Bridge.ino
@@ -63,10 +63,10 @@ IPAddress remoteIp;
 
 
 uint8_t buf1[bufferSize];
-uint8_t i1=0;
+uint16_t i1=0;
 
 uint8_t buf2[bufferSize];
-uint8_t i2=0;
+uint16_t i2=0;
